Moves Indiv to Domain
