+++ src/imageWriter.h
@@ -8,7 +8,7 @@
 #include <mutex>
 #include <condition_variable>
 #include <thread>
-#include "indiv.h"
+#include "domain/include/indiv.h"
 #include "params.h"
 #include "peeps.h"
